+++ Anchor/ArchiveView.swift
@@ -38,7 +38,7 @@ struct ArchiveView: View {
             }
             .padding(.horizontal, AnkerSpacing.screenPadding)
 #if os(iOS)
-            .padding(.bottom, AnkerSpacing.bottomBarClearance)
+            .padding(.bottom, AnkerSpacing.s5)
 #else
             .padding(.bottom, AnkerSpacing.s5)
 #endif